+++ GameObjects/Obstacles/Collectable.cpp
@@ -1,6 +1,6 @@
 #include "Collectable.h"
 
-Collectable::Collectable(float radius, string L,Vector2D Center): Center(Center) {
+Collectable::Collectable(double radius, string L,Vector2D Center): Center(Center) {
     this->radius=radius;
     this->L=L;
 
@@ -11,7 +11,7 @@ void Collectable::draw(Interface &interface)
     interface.drawCollectable(radius,L,Center);
 }
 
-Vector2D Collectable::collideWith(Ball &ball, float collision_time, Manager &manager)
+Vector2D Collectable::collideWith(Ball &ball, double collision_time, Manager &manager)
 {
     return {0,0};
 }
