+++ GameObjects/Obstacles/Collectable.h
@@ -4,13 +4,13 @@
 class Collectable:public Obstacle
 {
 private:
-    float radius;
+    double radius;
     string L;
     Vector2D Center;
 public:
-    Collectable(float radius,string L,Vector2D Center);
+    Collectable(double radius,string L,Vector2D Center);
     void draw(Interface & interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager & manager) override;
 };
