+++ GameObjects/Obstacles/InternalFrames.h
@@ -8,16 +8,16 @@ class InternalFrames: public Obstacle
 private:
     bool is45;
     bool isLeft;
-    float Diameter;
-    float LineXCoordinate;
-    float LineYCoordinate;
-    float InclinationAngle;
-    float setPositionX;
-    float setPositionYRation;
+    double Diameter;
+    double LineXCoordinate;
+    double LineYCoordinate;
+    double InclinationAngle;
+    double setPositionX;
+    double setPositionYRation;
 
 public:
-    explicit InternalFrames(bool is45,bool isLeft,float Diameter, float LineXCoordinate, float LineYCoordinate, float InclinationAngle, float setPositionX,float setPositionYRation);
+    explicit InternalFrames(bool is45,bool isLeft,double Diameter, double LineXCoordinate, double LineYCoordinate, double InclinationAngle, double setPositionX,double setPositionYRation);
     void draw(Interface & interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager & manager) override;
 };
