+++ GameObjects/Obstacles/PopBumper.h
@@ -6,9 +6,9 @@ class PopBumper : public Bumper{
 private:
 
 public:
-    PopBumper(Vector2D pos, float rad);
+    PopBumper(Vector2D pos, double rad);
     void draw(Interface & interface) override;
-    Vector2D collideWith(Ball& ball, float collision_time, Manager & manager)override;
+    Vector2D collideWith(Ball& ball, double collision_time, Manager & manager)override;
 };
