+++ GameObjects/Obstacles/Portal.cpp
@@ -6,7 +6,7 @@ void Portal::draw(Interface &interface) {
 
 Portal::Portal(Vector2D Pos1, Vector2D Pos2, double radius): Pos1(Pos1),Pos2(Pos2),radius(radius) {}
 
-Vector2D Portal::collideWith(Ball &ball, float collision_time, Manager &manager) {
+Vector2D Portal::collideWith(Ball &ball, double collision_time, Manager &manager) {
     bool Check1,Check2;
     Check1=BICCollision(ball,Pos1,radius);
     Check2=BICCollision(ball,Pos2,radius);
