+++ GameObjects/Obstacles/Wall.h
@@ -12,7 +12,7 @@ private:
 public:
     explicit Wall(float position,bool vertical);
     void draw(Interface & interface) override;
-    Vector2D collideWith(Ball & ball, float collision_time, Manager & manager) override;
+    Vector2D collideWith(Ball & ball, double collision_time, Manager & manager) override;
 };
